+++ src/treeview_switches.ml
@@ -39,7 +39,7 @@ module Dependency = struct
     match dependency with
     | Switch _ ->
       let selected = if is_current_sandbox then "-selected" else "" in
-      TreeItem.LightDarkIcon.
+      LightDarkIcon.
         { light =
             `String
               (Path.asset @@ "dependency-light" ^ selected ^ ".svg"
@@ -50,7 +50,7 @@ module Dependency = struct
               |> Path.to_string)
         }
     | Package _ ->
-      TreeItem.LightDarkIcon.
+      LightDarkIcon.
         { light = `String (Path.asset "number-light.svg" |> Path.to_string)
         ; dark = `String (Path.asset "number-dark.svg" |> Path.to_string)
         }